Ugly, semi-effective solution to dealing with addressElement's City UITextField's issues.
For some reason, for that SPECIFIC field, springs and struts not operating as normal.  I attempted a 'nuclear' solution by deleting the field and replacing it with a new one, and that failed suggesting that it isn't an XIB corruption issue.  I was unable to find any place in code where the field's position was being manipulated, so I have no clue what the real cause is.